波形ファイルの読み込みのコードを分離
