+++ SFZ_MIDI_Player/SFZ_MIDI_Player.vcxproj.filters
@@ -159,6 +159,9 @@
     <ClCompile Include="source\SamplePlayer.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="source\AudioLoader.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Image Include="App\icon.ico">
@@ -810,6 +813,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="include\PianoRoll.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="include\AudioLoader.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
