+++ SFZ_MIDI_Player/source/SamplePlayer.cpp
@@ -5,6 +5,7 @@
 #include <PianoRoll.hpp>
 #include <MIDILoader.hpp>
 #include <SampleSource.hpp>
+#include <AudioLoader.hpp>
 
 namespace
 {
